+++ lib/open_api/dsl/api_info_obj.rb
@@ -57,12 +57,12 @@ module OpenApi
       #         :export! => { type: Boolean }
       #   }
       %i[ header header! path path! query query! cookie cookie! ].each do |param_type|
-        define_method "do_#{param_type}" do |by:|
-          by.each do |key, value|
+        define_method "do_#{param_type}" do |by:, **common_schema|
+          by.each do |p_name, schema|
             action = param_type.to_s.delete('!')
-            type, value = value.is_a?(Hash) ? [value[:type], value] : [value, { }]
-            args = [ key.to_s.delete('!').to_sym, type, value ]
-            param_type['!'] || key['!'] ? send("#{action}!", *args) : send(action, *args)
+            type, schema = schema.is_a?(Hash) ? [schema[:type], schema] : [schema, { }]
+            args = [ p_name.to_s.delete('!').to_sym, type, schema.reverse_merge!(common_schema) ]
+            param_type['!'] || p_name['!'] ? send("#{action}!", *args) : send(action, *args)
           end
         end
       end